refactor(remediations): promote REMEDIATIONS to src/ + guidance-only gate
Move REMEDIATIONS/remediationsFor from ci/onboarding-harness to src/ so harness,
in-app Fix, and installer share one source. Add GUIDANCE_ONLY + autoFixCommandFor
gate (tailscale: installable but never clears check). No behavior change.
